Move water mesh + simple state generation to components library.
